Migrated to slim ~> 1.3.0
