+++ lib/skim.rb
@@ -2,7 +2,8 @@ require "temple"
 require "temple/coffee_script"
 
 require "slim"
-require "skim/compiler"
+require "skim/control_structures"
+require "skim/code_attributes"
 require "skim/sections"
 require "skim/engine"
 require "skim/template"
