+++ skim.gemspec
@@ -17,8 +17,8 @@ Sprockets-based asset pipeline.}
   gem.require_paths = ["lib"]
   gem.version       = Skim::VERSION
 
-  gem.add_dependency "slim", "<= 1.2.2"
-  gem.add_dependency "temple", "<= 0.4.1"
+  gem.add_dependency "slim", "~> 1.3.0"
+  gem.add_dependency "temple", "~> 0.4.0"
   gem.add_dependency "coffee-script"
   gem.add_dependency "coffee-script-source", ">= 1.2.0"
   gem.add_dependency "multi_json"
